+++ src/rovr/app.py
@@ -50,7 +50,7 @@ from rovr.action_buttons import (
     ZipButton,
 )
 from rovr.action_buttons.sort_order import SortOrderButton
-from rovr.classes.app_mixins import DragAndDrop, ThemeHandler
+from rovr.classes.app_mixins import DragAndDrop, KeyHandler, ThemeHandler
 from rovr.classes.mixins import Action, Actionable
 from rovr.classes.theme import RovrStylesheet
 from rovr.classes.type_aliases import KeysConfig
@@ -104,7 +104,12 @@ if constants.SCREENSHOT_LOCATION:
 
 
 class Application(
-    Actionable, ThemeHandler, DragAndDrop, DNDApp, inherit_bindings=False
+    Actionable,
+    ThemeHandler,
+    DragAndDrop,
+    KeyHandler,
+    DNDApp,
+    inherit_bindings=False,
 ):
     # our own form of BINDINGS that utilises check_key
     # key: str the action to use
@@ -885,79 +890,8 @@ class Application(
         console.print(screen_render)
         return console.export_svg(title="")
 
-    @staticmethod
-    def shorten_key(key: str) -> str:
-        from textual.keys import key_to_character
-
-        *modifiers, name = key.split("+")
-        character = key_to_character(name)
-
-        if character is not None and character.isprintable() and character != " ":
-            name = character
-
-        return "+".join((*modifiers, name))
-
-    async def _check_bindings(self, key: str, priority: bool = False) -> bool:
-        if not self.keys:
-            return await super()._check_bindings(key, priority)
-        self.notify(str(self.focused.check_consume_key(key, self.shorten_key(key))))
-        if (
-            priority
-            and self.focused is not None
-            and self.focused.check_consume_key(key, self.shorten_key(key))
-        ):
-            return False
-
-        namespaces = self._key_namespaces()
-        contexts = [("global", self)] if priority else self._active_key_contexts()
-        for context, namespace in contexts:
-            context = self.keys.get(context, {})
-            binding = context.get(key)
-            if isinstance(binding, dict):
-                action = binding["action"]
-            elif isinstance(binding, str):
-                action = binding
-            else:
-                action = None
-            if action == "noop":
-                return True
-            if action is not None and await self.run_action(
-                action,
-                default_namespace=namespace,
-                namespaces=namespaces,
-            ):
-                return True
-        return False
-
-    def _active_key_contexts(self) -> list[tuple[str, DOMNode]]:
-        contexts: list[tuple[str, DOMNode]] = []
-        focused = self.focused
-        nodes = focused.ancestors_with_self if focused is not None else [self.screen]
-        for node in nodes:
-            contexts.extend(
-                (context, node) for context in getattr(node, "key_contexts", ())
-            )
-            if node is self.screen:
-                break
-        if len(self.screen_stack) == 1:
-            contexts.append(("main", self))
-        return contexts
-
-    def _key_namespaces(self) -> dict[str, DOMNode]:
-        namespaces: dict[str, DOMNode] = {"app": self, "screen": self.screen}
-        for node in self.screen.walk_children(with_self=True):
-            contexts = getattr(node, "key_contexts", ())
-            if contexts:
-                namespaces.setdefault(contexts[0], node)
-
-        if self.focused is not None:
-            for node in reversed(self.focused.ancestors_with_self):
-                contexts = getattr(node, "key_contexts", ())
-                if contexts:
-                    namespaces[contexts[0]] = node
-        return namespaces
-
     def get_system_commands(self, screen: Screen) -> Iterable[SystemCommand]:
+        # TODO: remove command palette for a custom version some day
         yield SystemCommand(
             "Change theme",
             "Change the current theme",
